requirements.txt
astroid==4.0.1
colorama==0.4.6
dill==0.4.0
isort==7.0.0
mccabe==0.7.0
numpy==2.4.4
platformdirs==4.5.0
pygame==2.6.1
pygame-ce==2.5.7
pygame_gui==0.6.14
pylint==4.0.2
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
python-i18n==0.3.9
PyTMX==3.32
tomlkit==0.13.3
